chore: CheckBoxButton - Image fileprivate extension 추가
